pyproject.toml
[project]
name = "financial-documents-clustering"
description = "financial documents clustering"
version = "1.0"
dependencies = [
    "pip-tools",
    "pre-commit",
    "pytest",
    "kagglehub",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "python-dotenv",
    "click",
    "pydantic",
    "pytest",
    "scikit-learn"
]